+++ CHANGES.rst
@@ -27,6 +27,7 @@ UNRELEASED
  * **[FEATURE]** Support building wheel binary distributions.
  * **[CHANGE]** The dependency on ``update_checker`` has been increased
    to >= 0.8.
+ * **[REDDIT]** Add support for changes to UserLists on reddit.
  * **[BUGFIX]** Fix configuration parsing for ``store_json_result``.
  * **[BUGFIX]** Fix duplicate bug in :class:`.BoundedSet`.